ollama app2/requirements.txt
langchain_core
streamlit
langchain_community
python-dotenv